server moved to main.py and dockerfile changes
