--- core-management/READ_ME.md
@@ -1,6 +1,0 @@
-CURRENTLY RUNS SERVER AND CLIENT (which both close after API function call)
-
-docker build -t your_image_name .
-docker run --rm your_image_name
-
-
