+++ core_management/.gitignore
