+++ core_management/READ_ME.md
@@ -1,0 +1,4 @@
+CURRENTLY RUNS SERVER AND TEST CASES (Using mock values)
+
+docker build -t osiris-function-library -f core_management/Dockerfile .docker build -t your_image_name .
+docker run -it osiris-function-library
