+++ core_management/proto/core_management.proto
