+++ core_management/server/util.py
@@ -1,5 +1,4 @@
-# Placeholder database
-function_library = {}
+from server.data_store import function_library
 
 def addFunctionToLibrary(function_name: str, code: str, runtime: str, version: str = "1.0") -> bool:
     # Create a key for the database entry
