+++ core_management/team-10.md
